--- FTP/src/FTPConnectAndLoginDemo.java
@@ -1,43 +1,0 @@
-import java.io.IOException;
-import org.apache.commons.net.ftp.FTPClient;
-import org.apache.commons.net.ftp.FTPReply;
-public class FTPConnectAndLoginDemo {
-    private static void showServerReply(FTPClient ftpClient) {
-        String[] replies = ftpClient.getReplyStrings();
-        if (replies != null && replies.length > 0) {
-            for (String aReply : replies) {
-                System.out.println("SERVER: " + aReply);
-            }
-        }
-    }
-    public static void main(String[] args) {
-        String server = args[0];//"193.219.28.2";
-        int port = Integer.parseInt(args[1]); //21;
-        String user = args[2];//"anonymous";
-        String pass = args[3]; //"me@nowhere.com";
-        FTPClient ftpClient = new FTPClient();
-        try {
-            ftpClient.connect(server, port);
-            showServerReply(ftpClient);
-            int replyCode = ftpClient.getReplyCode();
-
-            // FTPReply stores a set of constants for FTP reply codes. To interpret the meaning of the codes
-            if (!FTPReply.isPositiveCompletion(replyCode)) { // interpret the reply code, return true if succesfully connected
-                System.out.println("Operation failed. Server reply code: " + replyCode);
-                return;
-            }
-            boolean success = ftpClient.login(user, pass);
-            showServerReply(ftpClient);
-            if (!success) {
-                System.out.println("Could not login to the server");
-                return;
-            } else {
-                System.out.println("LOGGED IN SERVER");
-            }
-        } catch (IOException ex) {
-            System.out.println("Oops! Something wrong happened");
-            ex.printStackTrace();
-        }
-    }
-}
-
